// src/app/@core/models/user.model.ts
import { Country } from './country.model';

export class User {
    _id?: number;
    firstName: string;
    lastName: string;
    email?: string;
    mobileNo?: string;
    isBanUser?: string;
    companyName?: string;
    userType?: any;
    profilePhoto?: any;
	uniqueCode?: any;
    companyID?: any;

}